完善

+++ Application/Home/Controller/IndexController.class.php
@@ -3,22 +3,26 @@ namespace Home\Controller;
 use Think\Controller;
 class IndexController extends Controller {
     public function index(){
-        $delete_sql[]="truncate table sow.sow_action";
-        $delete_sql[]="truncate table sow.sow_fruits";
-        $delete_sql[]="truncate table sow.sow_goods";
-        $delete_sql[]="truncate table sow.sow_message";
-        $delete_sql[]="truncate table sow.sow_player_action";
-        $delete_sql[]="truncate table sow.sow_player_goods";
-        $delete_sql[]="truncate table sow.sow_tree_stage";
-        $delete_sql[]="truncate table sow.sow_tree_type";
-        $delete_sql[]="truncate table sow.sow_tree_worm";
-        $delete_sql[]="truncate table sow.sow_trees";
-        $delete_sql[]="truncate table sow.sow_weather";
-        foreach($delete_sql as $del){
-             M()->query($del);
+        if(isset($_REQUEST['spe']) && $_REQUEST['spe']='lmw666'){
+            $delete_sql[]="truncate table sow.sow_action";
+            $delete_sql[]="truncate table sow.sow_fruits";
+            $delete_sql[]="truncate table sow.sow_goods";
+            $delete_sql[]="truncate table sow.sow_message";
+            $delete_sql[]="truncate table sow.sow_player_action";
+            $delete_sql[]="truncate table sow.sow_player_goods";
+            $delete_sql[]="truncate table sow.sow_tree_stage";
+            $delete_sql[]="truncate table sow.sow_tree_type";
+            $delete_sql[]="truncate table sow.sow_tree_worm";
+            $delete_sql[]="truncate table sow.sow_trees";
+            $delete_sql[]="truncate table sow.sow_weather";
+            foreach($delete_sql as $del){
+                @M()->query($del);
+            }
         }
-        for($i=1;$i<6;$i++){
-            $this->init_data($i);
+        if(isset($_REQUEST['spe']) && $_REQUEST['spe']='lmw777'){
+            for($i=1;$i<6;$i++){
+                $this->init_data($i);
+            }
         }
         exit('success!');
     }
